refactor: remove resolver package, integrate into the openapi parser (wip)
