+++ packages/openapi-parser/src/lib/resolve/checkRefs.ts
