+++ packages/openapi-parser/src/lib/resolve/escapeJsonPointer.ts
