+++ packages/openapi-parser/src/lib/resolve/isObject.ts
