+++ packages/openapi-parser/src/lib/resolve/replaceRefs.ts
