+++ packages/openapi-parser/src/lib/resolve/resolveRefs.test.ts
