+++ packages/openapi-parser/src/lib/resolve/resolveRefs.ts
