+++ packages/openapi-parser/src/lib/resolve/unescapeJsonPointer.ts
