--- packages/resolver/README.md
@@ -1,3 +1,0 @@
-# Scalar Resolver
-
-WIP
