--- packages/resolver/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@scalar/resolver",
-  "description": "Resolve JSON Schema and OpenAPI references",
-  "keywords": [
-    "openapi",
-    "scalar",
-    "swagger",
-    "json-schema",
-    "resolver",
-    "typescript"
-  ],
-  "license": "MIT",
-  "author": "Scalar (https://github.com/scalar)",
-  "homepage": "https://github.com/scalar/scalar",
-  "bugs": "https://github.com/scalar/scalar/issues/new/choose",
-  "version": "0.2.0",
-  "engines": {
-    "node": ">=18"
-  },
-  "type": "module",
-  "files": [
-    "dist",
-    "CHANGELOG.md"
-  ],
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "module": "./dist/index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/scalar/openapi-parser.git",
-    "directory": "packages/resolver"
-  },
-  "scripts": {
-    "dev": "vite dev",
-    "types:check": "tsc --noEmit --skipLibCheck",
-    "build": "pnpm types:check && vite build",
-    "test": "vitest"
-  },
-  "devDependencies": {
-    "vite-plugin-dts": "^3.7.3"
-  },
-  "dependencies": {
-    "@types/node": "^20.11.20",
-    "vite": "^5.1.4"
-  }
-}
