--- packages/resolver/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "tsBuildInfoFile": null,
-    "composite": true,
-    "skipLibCheck": true,
-    "resolveJsonModule": true,
-    "moduleResolution": "Node",
-    "esModuleInterop": true,
-    "module": "esnext",
-    "allowSyntheticDefaultImports": true,
-    "outDir": "dist",
-    "sourceMap": false,
-    "lib": ["ESNext", "DOM"],
-    "target": "ESNext",
-    "types": ["@modyfi/vite-plugin-yaml/modules"]
-  },
-  "include": ["src/**/*", "tests/**/*", "**/*.json", "**/*.yaml"]
-}
